+++ ReleaseNotes.md
@@ -116,3 +116,5 @@ Please change your configuration according to [App.xaml](https://github.com/spie
 * New `SaveFileControl.ForceFileExtensionOfFileFilter` property to enforce a file extension of the selected filter
 * Better support of dark theme in file system dialogs
 * Localization for Russian
+#### Fixes
+* Fixed `DateTime` conversion inside `DateTimeAgoConverter`